@@ -11,9 +11,11 @@
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.7.0",
     "react-scripts": "4.0.3",
+    "react-stl-viewer": "^2.2.4",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
